+++ src/main/kotlin/cloud/glitchdev/rfu/utils/rendering/Render3D.kt
@@ -28,12 +28,12 @@ object Render3D {
         location: Vec3,
         radius: Float,
         color: Color,
+        sliceColor: Color? = null,
         context: WorldRenderContext,
         stacks: Int = 16,
         slices: Int = 16,
         lineWidth: Float = 2.0f,
-        filled: Boolean = false,
-        drawSlices: Boolean = true
+        filled: Boolean = false
     ) {
         if (!isVisible(buildSphereBounds(location, radius))) {
             return
@@ -93,7 +93,7 @@ object Render3D {
             }
         }
 
-        if (drawSlices) {
+        if (sliceColor != null) {
             //? if >=1.21.11 {
             val buffer = consumers.getBuffer(RenderTypes.LINES)
             //?} else {
@@ -118,11 +118,11 @@ object Render3D {
                     val x1 = cos(lng1)
                     val y1 = sin(lng1)
 
-                    drawVertex(buffer, matrix, (x0 * zr0).toFloat(), z0.toFloat(), (y0 * zr0).toFloat(), color, lineWidth)
-                    drawVertex(buffer, matrix, (x0 * zr1).toFloat(), z1.toFloat(), (y0 * zr1).toFloat(), color, lineWidth)
+                    drawVertex(buffer, matrix, (x0 * zr0).toFloat(), z0.toFloat(), (y0 * zr0).toFloat(), sliceColor, lineWidth)
+                    drawVertex(buffer, matrix, (x0 * zr1).toFloat(), z1.toFloat(), (y0 * zr1).toFloat(), sliceColor, lineWidth)
 
-                    drawVertex(buffer, matrix, (x0 * zr1).toFloat(), z1.toFloat(), (y0 * zr1).toFloat(), color, lineWidth)
-                    drawVertex(buffer, matrix, (x1 * zr1).toFloat(), z1.toFloat(), (y1 * zr1).toFloat(), color, lineWidth)
+                    drawVertex(buffer, matrix, (x0 * zr1).toFloat(), z1.toFloat(), (y0 * zr1).toFloat(), sliceColor, lineWidth)
+                    drawVertex(buffer, matrix, (x1 * zr1).toFloat(), z1.toFloat(), (y1 * zr1).toFloat(), sliceColor, lineWidth)
                 }
             }
         }